backend/pyproject.toml
[project]
name = "tpe-cloud-2025"
version = "0.1.0"
description = "Sistema de compras mayoristas comunitarias en la nube"
requires-python = ">=3.12.9"
readme = "README.md"
dependencies = [
    "boto3>=1.34.162",
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "django-storages[boto3]>=1.14.6",
    "djangorestframework>=3.16.1",
    "drf-nested-routers>=0.94.2",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.9",
]

[tool.black]
line-length = 140
